step_05/testbedEVSE/requirements.txt
fastapi
uvicorn
pydantic
websockets
aiosqlite
pandas
ocpp
python-dotenv